requirements.txt
pytorch==2.0.1
matplotlib==3.7.1
seaborn==0.12.2
datasets==2.12.0
transformers==4.29.2
